+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 1.8.0
+* Updated to 26.2
+* [Client]
+    * Fix realms not being displayed as expected
+    * Reposition configuration button in Online settings
+    * New advanced options
+        * Integrated server: Disable `enforce-secure-profile` by default
+        * Compact the title screen (reposition elements / remove redundant buttons)
+        * "Lock down"
+    * Skip friendlist promotion
+* Removed legacy configuration directory migration
+* Minor performance improvements and tweaks
+
 # 1.7.0
 * [Client] Added
     * `force-secure-skin-download`: Forces the skin to be downloaded using a secure (https) connection